Refactor routing of images
